DataBase updated for Proliferation experiment storage

+++ src/main/java/com/biotools/repository/UserRepository.java
@@ -7,8 +7,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.biotools.entity.User;
 
-public interface UserRepository extends JpaRepository<User,Integer> {
-	User findUserById(Integer id);
+public interface UserRepository extends JpaRepository<User,Long> {
+	User findUserById(Long id);
 	
 	Optional<User> findUserByUsername(String username);
